+++ python/hidet/ffi/__init__.py
@@ -9,10 +9,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .ffi import _LIB, _LIB_NCCL
-from .runtime_api import runtime_api, nccl_available
-if nccl_available():
-    from .runtime_api import nccl_runtime_api, NcclUniqueId, NcclCommunicator, get_nccl_comm
+from .ffi import _LIB
+from .runtime_api import runtime_api
 
 from . import callbacks
 from . import crt
