+++ python/hidet/ffi/ffi.py
@@ -12,16 +12,13 @@
 from typing import List, Dict, Optional
 import os
 import os.path
-import glob
 import ctypes
-from hidet.libinfo import get_library_search_dirs, get_nccl_library_search_dirs
+from hidet.libinfo import get_library_search_dirs
 
 _LIB: Optional[ctypes.CDLL] = None
 _LIB_RUNTIME: Optional[ctypes.CDLL] = None
-_LIB_NCCL: Optional[ctypes.CDLL] = None
 
-
-library_paths: Dict[str, Optional[str]] = {'hidet': None, 'hidet_runtime': None, 'nccl': None}
+library_paths: Dict[str, Optional[str]] = {'hidet': None, 'hidet_runtime': None}
 
 
 def load_library():
@@ -42,18 +39,7 @@ def load_library():
     if _LIB is None:
         raise OSError('Can not find library in the following directory: \n' + '\n'.join(library_dirs))
 
-def load_nccl_library():
-    global _LIB_NCCL
-    library_dirs = get_nccl_library_search_dirs()
-    for library_dir in library_dirs:
-        lib_nccl_paths = glob.glob(os.path.join(library_dir, 'libnccl.so*'))
-        if len(lib_nccl_paths) == 0:
-            continue
-        _LIB_NCCL = ctypes.cdll.LoadLibrary(lib_nccl_paths[0])
-        library_paths['nccl'] = lib_nccl_paths[0]
-        break
-    if _LIB_NCCL is None:
-        raise OSError('Can not find nccl library in the following directory: \n' + '\n'.join(library_dirs))
+
 
 def get_last_error() -> Optional[str]:
     func = getattr(get_last_error, '_func', None)
@@ -80,17 +66,17 @@ def func_exists(func_name: str, shared_lib: ctypes.CDLL) -> bool:
         return False
 
 
-def get_func(func_name, arg_types: List, restype):
+def get_func(func_name, arg_types: List, restype, lib=None):
     if func_exists(func_name, _LIB):
         func = getattr(_LIB, func_name)
     elif func_exists(func_name, _LIB_RUNTIME):
         func = getattr(_LIB_RUNTIME, func_name)
-    elif func_exists(func_name, _LIB_NCCL):
-        func = getattr(_LIB_NCCL, func_name)
+    elif func_exists(func_name, lib):
+        func = getattr(lib, func_name)
     else:
         raise ValueError(
-            'Can not find function "{}" in hidet libraries:\n{}\n{}\n{}'.format(
-                func_name, library_paths['hidet'], library_paths['hidet_runtime'], library_paths['nccl']
+            'Can not find function "{}" in hidet libraries:\n{}\n{}'.format(
+                func_name, library_paths['hidet'], library_paths['hidet_runtime']
             )
         )
 
@@ -112,7 +98,4 @@ def get_func(func_name, arg_types: List, restype):
 
 
 load_library()
-load_nccl_library()
 
-def nccl_available():
-    return _LIB_NCCL is not None
